ci(release.yaml): update GitHub Actions workflow to use latest versions and improve build process
The GitHub Actions workflow in `release.yaml` is updated to use the
latest versions of Ubuntu, checkout, setup-go, and goreleaser-action.
This ensures compatibility with the latest features and security
updates. The `go-version` is updated to 1.22 to leverage improvements
and new features in the latest Go release. Additionally, the `--clean`
argument is added to the goreleaser command to ensure a clean build
environment, reducing potential issues from previous builds.