feat: upgrade go-github to v62 and go-querystring to v1.1.0
This update upgrades the `go-github` library from v32 to v62 and the
`go-querystring` library from v1.0.0 to v1.1.0 across various internal
applications and dependencies. The upgrade to `go-github` v62 ensures
compatibility with the latest GitHub API changes and features, enhancing
the stability and capabilities of our GitHub integrations. The update to
`go-querystring` v1.1.0 introduces improvements and bug fixes that
enhance query string handling in HTTP requests.