feat(Dockerfile): update base image to golang:1.22-bullseye and simplify binary compression step
Updating the base image from golang:1.13.15-buster to
golang:1.22-bullseye ensures the application is built with the latest
features and security updates of the Go language and Debian Bullseye.
Additionally, the binary compression step is simplified by removing the
`strip` command and directly using `upx` for compression, which
streamlines the build process and potentially reduces the Docker image
size.