ci(ci.yaml): remove pull_request trigger, add docker build step after tests
The pull_request trigger is removed to streamline the CI process,
ensuring that actions are only triggered on pushes to the 'develop'
branch. This change reduces unnecessary builds on pull requests,
focusing resources on integration after code merges.
Additionally, a new job 'call-docker-build' is added to build a Docker
image using the 'docker-publish.yaml' workflow from the
'dictyBase/workflows' repository. This step is dependent on the
successful completion of the 'test' job, ensuring that Docker images are
only built when tests pass. This integration enhances the CI pipeline by
automating Docker image creation for successful builds, facilitating
continuous deployment practices.