refactor: downgrade go-github library from v62 to v32 for compatibility
The downgrade of the go-github library from version 62 to 32 across
multiple files ensures compatibility with other project dependencies
that require an older version of the library. Additionally, in
`migrate.go`, the introduction of `GetLegacyGithubClient` instead of the
standard client supports systems that rely on older GitHub API
functionalities, enhancing backward compatibility.